backend/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
gunicorn
django-cors-headers
djangorestframework
django-filter
